refactor: 게시물 리스트 api - useQuery 변경
